Major update
config.txt : New customizations for the main clock and the secondary clock, as well as for the news section & weather section.
DigitalWorld.hws : Small adjustments
smartclock.hws : Several fixed and improvements, the secondary clock is almost completed, you can return to the primary clock clicking on the clock are. Added many comments in the source code.